Refactor PI nav: add Overview tab, update routing
- Introduce "Overview" as default tab for active/completed PIs
- Split "Overview" and "Details" into separate pages
- Update PlanningIntervalNav and tests for new tab structure
- Enhance PlanningIntervalSwitcher to preserve sub-paths on switch
- Root PI page now redirects to correct tab based on PI state
- Update tests for new navigation and routing behaviors
