+++ Wayd.Web/src/wayd.web.reactclient/src/app/planning/planning-intervals/[key]/_components/planning-interval-switcher.test.tsx
@@ -25,9 +25,11 @@ Object.defineProperty(window, 'matchMedia', {
 })
 
 const mockPush = jest.fn()
+const mockUsePathname = jest.fn()
 
 jest.mock('next/navigation', () => ({
   useRouter: () => ({ push: mockPush }),
+  usePathname: () => mockUsePathname(),
 }))
 
 jest.mock('@/src/components/contexts/theme', () => ({
@@ -52,6 +54,7 @@ const mockQuery = useGetPlanningIntervalsQuery as unknown as jest.Mock
 describe('PlanningIntervalSwitcher', () => {
   beforeEach(() => {
     jest.clearAllMocks()
+    mockUsePathname.mockReturnValue('/planning/planning-intervals/2')
   })
 
   it('renders the trigger button even before the list has loaded', () => {
@@ -117,6 +120,62 @@ describe('PlanningIntervalSwitcher', () => {
     render(<PlanningIntervalSwitcher piKey={2} />)
     await user.click(screen.getByRole('button'))
     await user.click(await screen.findByText('2025 PI 1'))
+
+    expect(mockPush).toHaveBeenCalledWith('/planning/planning-intervals/1')
+  })
+
+  const clickSwitchTo = async (target: string) => {
+    mockQuery.mockReturnValue({
+      data: [
+        { key: 1, name: '2025 PI 1', start: '2025-01-01', state: { name: 'Completed' } },
+        { key: 2, name: '2026 PI 1', start: '2026-01-01', state: { name: 'Active' } },
+      ],
+    })
+    const user = userEvent.setup()
+
+    render(<PlanningIntervalSwitcher piKey={2} />)
+    await user.click(screen.getByRole('button'))
+    await user.click(await screen.findByText(target))
+  }
+
+  it.each([
+    ['/planning/planning-intervals/2/overview', '/planning/planning-intervals/1/overview'],
+    ['/planning/planning-intervals/2/details', '/planning/planning-intervals/1/details'],
+    ['/planning/planning-intervals/2/plan-review', '/planning/planning-intervals/1/plan-review'],
+    ['/planning/planning-intervals/2/objectives', '/planning/planning-intervals/1/objectives'],
+    ['/planning/planning-intervals/2/risks', '/planning/planning-intervals/1/risks'],
+    [
+      '/planning/planning-intervals/2/objectives/health-report',
+      '/planning/planning-intervals/1/objectives/health-report',
+    ],
+  ])('preserves the top-level sub-path when switching (%s)', async (pathname, expected) => {
+    mockUsePathname.mockReturnValue(pathname)
+
+    await clickSwitchTo('2025 PI 1')
+
+    expect(mockPush).toHaveBeenCalledWith(expected)
+  })
+
+  it('falls back to the PI root when the sub-path is an unknown or deep detail route', async () => {
+    mockUsePathname.mockReturnValue(
+      '/planning/planning-intervals/2/objectives/abc-123',
+    )
+
+    await clickSwitchTo('2025 PI 1')
+
+    // "objectives" is preserved; the deep segment "abc-123" is dropped so we
+    // don't land on a non-existent page on the other PI.
+    expect(mockPush).toHaveBeenCalledWith(
+      '/planning/planning-intervals/1/objectives',
+    )
+  })
+
+  it('falls back to the PI root when on an unrecognized sub-route', async () => {
+    mockUsePathname.mockReturnValue(
+      '/planning/planning-intervals/2/something-else',
+    )
+
+    await clickSwitchTo('2025 PI 1')
 
     expect(mockPush).toHaveBeenCalledWith('/planning/planning-intervals/1')
   })
