+++ Wayd.Web/src/wayd.web.reactclient/src/app/planning/planning-intervals/[key]/_components/planning-interval-switcher.tsx
@@ -3,11 +3,33 @@
 import { IconMenu } from '@/src/components/common'
 import { useGetPlanningIntervalsQuery } from '@/src/store/features/planning/planning-interval-api'
 import { SwapOutlined } from '@ant-design/icons'
-import { useRouter } from 'next/navigation'
+import { usePathname, useRouter } from 'next/navigation'
 import { useState } from 'react'
+
+// Top-level tabs that are safe to carry across PIs. Deeper routes (e.g. a
+// specific objective or risk) belong to one PI's data and would 404 on the
+// next, so those fall back to the PI root.
+const PRESERVABLE_SUBPATHS = [
+  'overview',
+  'details',
+  'plan-review',
+  'objectives/health-report',
+  'objectives',
+  'risks',
+]
+
+const resolveSubPath = (pathname: string, piKey: number) => {
+  const prefix = `/planning/planning-intervals/${piKey}`
+  if (!pathname.startsWith(prefix)) return ''
+  const rest = pathname.slice(prefix.length).replace(/^\/+/, '')
+  return PRESERVABLE_SUBPATHS.find(
+    (sub) => rest === sub || rest.startsWith(`${sub}/`),
+  )
+}
 
 const PlanningIntervalSwitcher = ({ piKey }: { piKey: number }) => {
   const router = useRouter()
+  const pathname = usePathname()
   const [queryEnabled, setQueryEnabled] = useState(false)
   const { data: piListData } = useGetPlanningIntervalsQuery(undefined, {
     skip: !queryEnabled,
@@ -26,7 +48,11 @@ const PlanningIntervalSwitcher = ({ piKey }: { piKey: number }) => {
         }))
 
   const handlePIChange = (value: string | number) => {
-    router.push(`/planning/planning-intervals/${value}`)
+    const subPath = resolveSubPath(pathname ?? '', piKey)
+    const target = subPath
+      ? `/planning/planning-intervals/${value}/${subPath}`
+      : `/planning/planning-intervals/${value}`
+    router.push(target)
   }
 
   const handleOpenChange = (open: boolean) => {
