+++ Wayd.Web/src/wayd.web.reactclient/src/app/planning/planning-intervals/[key]/overview/page.tsx
@@ -1,0 +1,52 @@
+'use client'
+
+import PageTitle from '@/src/components/common/page-title'
+import { useDocumentTitle } from '@/src/hooks/use-document-title'
+import useAuth from '@/src/components/contexts/auth'
+import { authorizePage } from '@/src/components/hoc'
+import { notFound, usePathname } from 'next/navigation'
+import { useAppDispatch } from '@/src/hooks'
+import { setBreadcrumbTitle } from '@/src/store/breadcrumbs'
+import PlanningIntervalDetailsLoading from '../loading'
+import { use, useEffect } from 'react'
+import { useGetPlanningIntervalQuery } from '@/src/store/features/planning/planning-interval-api'
+
+const PlanningIntervalOverviewPage = (props: {
+  params: Promise<{ key: string }>
+}) => {
+  const { key } = use(props.params)
+  const piKey = Number(key)
+
+  const pathname = usePathname()
+  const dispatch = useAppDispatch()
+
+  const { data: planningIntervalData, isLoading } =
+    useGetPlanningIntervalQuery(+piKey)
+
+  useDocumentTitle(`${planningIntervalData?.name ?? piKey} - PI Overview`)
+
+  useEffect(() => {
+    planningIntervalData &&
+      dispatch(
+        setBreadcrumbTitle({ title: planningIntervalData.name, pathname }),
+      )
+  }, [dispatch, pathname, planningIntervalData])
+
+  if (isLoading) {
+    return <PlanningIntervalDetailsLoading />
+  }
+
+  if (!isLoading && !planningIntervalData) {
+    return notFound()
+  }
+
+  return <PageTitle title="PI Overview" />
+}
+
+const PlanningIntervalOverviewPageWithAuthorization = authorizePage(
+  PlanningIntervalOverviewPage,
+  'Permission',
+  'Permissions.PlanningIntervals.View',
+)
+
+export default PlanningIntervalOverviewPageWithAuthorization
